requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2
djangorestframework==3.14.0
kombu==5.3.4
prompt-toolkit==3.0.43
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
